Migrate DataFusion example to new engine impl
